refactor StartPage component
